chapter13: truncated cylinder. Now with caps!

+++ src/main/java/org/schakalacka/java/raytracing/scene/Material.java
@@ -5,7 +5,7 @@ import org.schakalacka.java.raytracing.geometry.patterns.Pattern;
 import org.schakalacka.java.raytracing.math.Tuple;
 
 public record Material(Color color, double ambient, double diffuse, double specular, double shininess,
-                       float reflectivity, double transparency, double refractiveIndex, Pattern pattern,
+                       double reflectivity, double transparency, double refractiveIndex, Pattern pattern,
                        boolean createsShadow) {
 
     public static MaterialBuilder newMaterial() {
@@ -74,7 +74,7 @@ public record Material(Color color, double ambient, double diffuse, double specu
         private double diffuse = 0.9;
         private double specular = 0.9;
         private double shininess = 200.0;
-        private float reflectivity = 0.0f;
+        private double reflectivity = 0.0f;
         private double transparency = 0.0;
         private double refractiveIndex = 1.0;
         private Pattern pattern;
@@ -105,7 +105,7 @@ public record Material(Color color, double ambient, double diffuse, double specu
             return this;
         }
 
-        public MaterialBuilder reflectivity(float reflectivity) {
+        public MaterialBuilder reflectivity(double reflectivity) {
             this.reflectivity = reflectivity;
             return this;
         }
